+++ src/main/java/com/rajan/llm_cost_aware_gateway/dataplane/models/LLMRequest.java
@@ -1,5 +1,6 @@
 package com.rajan.llm_cost_aware_gateway.dataplane.models;
 
+import com.rajan.llm_cost_aware_gateway.controlplane.models.Request;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -11,7 +12,7 @@ import java.util.UUID;
 public class LLMRequest {
     private String model;
     private String prompt;
-    private int maxTokens;
+    private long maxTokens;
     private double temperature;
     private UUID requestId;
     private Map<String, Object> metadata;
